pyproject.toml
[tool.poetry]
name = "mlops-zoomcamp-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.3.0"
pandas = "^2.0.3"
xgboost = "^1.7.6"
prefect = "^2.10.21"
optuna = "^3.2.0"
mlflow = "^2.5.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.24.0"
black = "^23.7.0"
pylint = "^2.17.4"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pylint."messages control"]
# Only show warnings with the listed confidence levels. Leave empty to show all.
# Valid levels: HIGH, CONTROL_FLOW, INFERENCE, INFERENCE_FAILURE, UNDEFINED.
confidence = ["HIGH", "CONTROL_FLOW", "INFERENCE", "INFERENCE_FAILURE", "UNDEFINED"]

disable = ["raw-checker-failed", "bad-inline-option", "locally-disabled", "file-ignored",
    "suppressed-message", "useless-suppression", "deprecated-pragma",
    "use-symbolic-message-instead", "missing-module-docstring",
    "missing-function-docstring", "invalid-name"]


[tool.isort]
profile = "black"
